split colors and read plotting into separate modules
