+++ shabam/colors.py
@@ -1,0 +1,13 @@
+
+COLORS = {
+    'A':   [0.0, 0.4, 0.0], # green
+    'C':   [0.0, 0.0, 1.0], # blue
+    'G':   [1.0, 0.6, 0.2], # orange
+    'T':   [1.0, 0.0, 0.0], # red
+    'M':   [0.9, 0.9, 0.9], # match, gray
+    'M_f': [0.3, 0.7, 1.0], # forward match, orange
+    'M_r': [1.0, 0.7, 0.3], # reverse match, blue
+    '-':   [0.9, 0.4, 0.8], # deletion, pink
+    'I':   [0.4, 0.1, 0.8], # insertion, purple
+    'N':   [0.0, 0.0, 1.0], # unknown, black
+}
